+++ src/NumArray.php
@@ -188,6 +188,20 @@ class NumArray
         return self::eye($axis);
     }
 
+    public static function arange(float $start, float $stop, float $step = 1.0): NumArray
+    {
+        if (($start > $stop && $step > 0) || ($start < $stop && $step < 0)) {
+            return new NumArray([]);
+        }
+        $range = $step === 1.0 ? range($start, $stop) : range($start, $stop, $step);
+        $end = end($range);
+        if (($step > 0 && $end >= $stop) || ($step < 0 && $end <= $stop)) {
+            array_pop($range);
+        }
+        reset($range);
+        return new NumArray($range);
+    }
+
     private static function recursiveToString(array $data, int $level = 0): string
     {
         $indent = str_repeat("  ", $level);